move all types to single mod, bump MSRV
