+++ Cargo.toml
@@ -4,7 +4,7 @@ version = "0.4.0"
 edition = "2021"
 authors = ["Michal Borejszo <michael.borejszo@gmail.com>"]
 license = "MIT"
-rust-version = "1.56"
+rust-version = "1.60"
 description = "Wavefrom generator library"
 repository = "https://github.com/spitfire05/wavegen-rs"
 readme = "README.md"
@@ -31,7 +31,7 @@ hound = "3.5.0"
 [features]
 default = ["std"]
 libm = ["dep:libm", "num-traits/libm"]
-std = []
+std = ["num-traits/std"]
 
 [lib]
 bench = false
